edr_server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
python-dateutil==2.8.2
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
apscheduler==3.10.4
dj-database-url==2.1.0
